lambdas/process_prompts/requirements.txt
requests
ai21
psycopg2-binary